Rename Resort folder
