+++ src/components/Pages/SplashScreen.js
